Type fixes Refactor settings code
