+++ js/util/global.d.ts
@@ -116,6 +116,7 @@ declare global {
 		V3_subtract(x: number, y: number, z: number): ArrayVector3
 		V3_subtract(values: ArrayVector3): ArrayVector3
 		V3_subtract(value: THREE.Vector3): ArrayVector3
+		V3_multiply(scalar: number): ArrayVector3
 		V3_multiply(x: number, y: number, z: number): ArrayVector3
 		V3_multiply(values: ArrayVector3): ArrayVector3
 		V3_multiply(value: THREE.Vector3): ArrayVector3
